+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: hcaspar <hcaspar@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/26 17:32:54 by hcaspar           #+#    #+#             */
-/*   Updated: 2017/04/05 20:33:19 by hcaspar          ###   ########.fr       */
+/*   Updated: 2017/04/18 14:44:10 by hcaspar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ t_env				*init_env(void)
 void				init_values(t_env *e)
 {
 	e->pu = 1;
-	e->oldtime = (double)clock();
+	e->oldtime = (float)clock();
 	e->zoom = 1.0;
 	e->v.x = -1.6;
 	e->v.y = -1.6;
